SpreadsheetExpressionReferenceLoader was SpreadsheetExpressionReferenceContext
